+++ packages/e2e/workflow/overlay-image-annotation.test.ts
@@ -1,0 +1,135 @@
+import { describe, it, expect, beforeAll, afterAll, vi } from 'vitest'
+import { prisma } from '@shumai/db'
+import { setupTestDbHooks } from '@shumai/db/test'
+import { workflowService, TaskQueueTranscode } from '@shumai/workflow-core'
+import { initTranscodeWorkflows } from '@shumai/transcode'
+import { s3Service } from '@shumai/core/src/s3/s3'
+import { fileURLToPath } from 'url'
+import * as path from 'path'
+import * as fs from 'fs'
+
+const currentDir = path.dirname(fileURLToPath(import.meta.url))
+const transcodeWorkflowsPath = path.resolve(currentDir, '../../../apps/transcode/src/workflows.ts')
+const fixturesDir = path.resolve(currentDir, '../fixtures')
+
+describe.each(['local', 'temporal'] as const)(
+  'Workflow E2E - overlayImageAnnotationWorkflow (executor: %s)',
+  (mode) => {
+    setupTestDbHooks()
+
+    let transcodeWorkerPromise: Promise<void> | null = null
+
+    beforeAll(async () => {
+      process.env.S3_BUCKET = 'shumai-e2e-test-bucket-transcode'
+
+      workflowService.setExecutorType(mode)
+      initTranscodeWorkflows()
+
+      if (mode === 'temporal') {
+        console.log('Starting background worker for transcode Temporal E2E tests...')
+        transcodeWorkerPromise = workflowService.startWorkers(TaskQueueTranscode, {
+          workflowsPath: transcodeWorkflowsPath,
+        })
+        await new Promise((resolve) => setTimeout(resolve, 2000))
+      } else {
+        console.log('Starting local workflow service polling...')
+        workflowService.start()
+      }
+    })
+
+    afterAll(async () => {
+      if (mode === 'temporal') {
+        console.log('Shutting down Temporal workers...')
+        await workflowService.shutdownWorkers()
+        await Promise.all([transcodeWorkerPromise].filter(Boolean))
+      }
+      workflowService.close()
+      vi.restoreAllMocks()
+      try {
+        console.log('Cleaning up local E2E storage files...')
+        await s3Service.deletePrefix('shumai-e2e-test-bucket-transcode', '')
+      } catch (err) {
+        console.error('Failed to clean up E2E storage folder:', err)
+      }
+    })
+
+    it('should run transcodeMedia image annotation overlay successfully', async () => {
+      // 1. Seed Database
+      const team = await prisma.team.create({
+        data: { name: 'E2E Image Annotation Team' },
+      })
+
+      const project = await prisma.project.create({
+        data: { name: 'E2E Image Annotation Project', teamId: team.id },
+      })
+
+      const storageKey = await prisma.storageKey.create({
+        data: {
+          key: 'projects/e2e/image-ann.png',
+        },
+      })
+
+      const asset = await prisma.asset.create({
+        data: {
+          name: 'image-ann.png',
+          type: 'file',
+          status: 'uploaded',
+          mediaType: 'image/png',
+          projectId: project.id,
+          storageKeyId: storageKey.id,
+        },
+      })
+
+      // 2. Seed S3 Storage from Fixture
+      const pngPath = path.join(fixturesDir, 'small.png')
+      const pngBuffer = fs.readFileSync(pngPath)
+      await s3Service.putObject(
+        'shumai-e2e-test-bucket-transcode',
+        'projects/e2e/image-ann.png',
+        pngBuffer,
+        pngBuffer.length,
+        'image/png',
+      )
+
+      // 3. Create Workflow Task
+      const task = await prisma.workflowTask.create({
+        data: {
+          type: 'transcode_image_annotation',
+          status: 'pending',
+          assetId: asset.id,
+          projectId: project.id,
+          teamId: team.id,
+          payload: {
+            projectId: project.id,
+            imageAnnotation: {
+              annotations: [
+                {
+                  type: 'box',
+                  color: '#ff0000',
+                  points: [
+                    [0.1, 0.1],
+                    [0.9, 0.9],
+                  ],
+                },
+              ],
+            },
+          },
+        },
+      })
+
+      // 4. Wait for workflow to complete
+      console.log(
+        `Submitted E2E Image Annotation Workflow Task. ID: ${task.id}. Awaiting completion...`,
+      )
+      const completedTask = await workflowService.executeWait(task, 45000)
+
+      // 5. Verification
+      expect(completedTask.status).toBe('completed')
+
+      // Verify task output contains overlay image key
+      const output = completedTask.output as unknown as { key: string }
+      expect(output).toBeDefined()
+      expect(output.key).toContain('annotations/')
+    }, 50000)
+  },
+)
