+++ packages/e2e/workflow/take-video-screenshots.test.ts
@@ -1,0 +1,130 @@
+import { describe, it, expect, beforeAll, afterAll, vi } from 'vitest'
+import { prisma } from '@shumai/db'
+import { setupTestDbHooks } from '@shumai/db/test'
+import { workflowService, TaskQueueTranscode } from '@shumai/workflow-core'
+import { initTranscodeWorkflows } from '@shumai/transcode'
+import { s3Service } from '@shumai/core/src/s3/s3'
+import { fileURLToPath } from 'url'
+import * as path from 'path'
+import * as fs from 'fs'
+
+const currentDir = path.dirname(fileURLToPath(import.meta.url))
+const transcodeWorkflowsPath = path.resolve(currentDir, '../../../apps/transcode/src/workflows.ts')
+const fixturesDir = path.resolve(currentDir, '../fixtures')
+
+describe.each(['local', 'temporal'] as const)(
+  'Workflow E2E - takeVideoScreenshotsWorkflow (executor: %s)',
+  (mode) => {
+    setupTestDbHooks()
+
+    let transcodeWorkerPromise: Promise<void> | null = null
+
+    beforeAll(async () => {
+      process.env.S3_BUCKET = 'shumai-e2e-test-bucket-transcode'
+
+      workflowService.setExecutorType(mode)
+      initTranscodeWorkflows()
+
+      if (mode === 'temporal') {
+        console.log('Starting background worker for transcode Temporal E2E tests...')
+        transcodeWorkerPromise = workflowService.startWorkers(TaskQueueTranscode, {
+          workflowsPath: transcodeWorkflowsPath,
+        })
+        await new Promise((resolve) => setTimeout(resolve, 2000))
+      } else {
+        console.log('Starting local workflow service polling...')
+        workflowService.start()
+      }
+    })
+
+    afterAll(async () => {
+      if (mode === 'temporal') {
+        console.log('Shutting down Temporal workers...')
+        await workflowService.shutdownWorkers()
+        await Promise.all([transcodeWorkerPromise].filter(Boolean))
+      }
+      workflowService.close()
+      vi.restoreAllMocks()
+      try {
+        console.log('Cleaning up local E2E storage files...')
+        await s3Service.deletePrefix('shumai-e2e-test-bucket-transcode', '')
+      } catch (err) {
+        console.error('Failed to clean up E2E storage folder:', err)
+      }
+    })
+
+    it('should run transcodeMedia screenshot extraction successfully', async () => {
+      // 1. Seed Database
+      const team = await prisma.team.create({
+        data: { name: 'E2E Video Screenshot Team' },
+      })
+
+      const project = await prisma.project.create({
+        data: { name: 'E2E Video Screenshot Project', teamId: team.id },
+      })
+
+      const storageKey = await prisma.storageKey.create({
+        data: {
+          key: 'projects/e2e/video-shot.mp4',
+        },
+      })
+
+      const asset = await prisma.asset.create({
+        data: {
+          name: 'video-shot.mp4',
+          type: 'file',
+          status: 'uploaded',
+          mediaType: 'video/mp4',
+          projectId: project.id,
+          storageKeyId: storageKey.id,
+        },
+      })
+
+      // 2. Seed S3 Storage from Fixture
+      const mp4Path = path.join(fixturesDir, 'small.mp4')
+      const mp4Buffer = fs.readFileSync(mp4Path)
+      await s3Service.putObject(
+        'shumai-e2e-test-bucket-transcode',
+        'projects/e2e/video-shot.mp4',
+        mp4Buffer,
+        mp4Buffer.length,
+        'video/mp4',
+      )
+
+      // 3. Create Workflow Task
+      const task = await prisma.workflowTask.create({
+        data: {
+          type: 'transcode_screenshot',
+          status: 'pending',
+          assetId: asset.id,
+          projectId: project.id,
+          teamId: team.id,
+          payload: {
+            projectId: project.id,
+            screenshot: {
+              start: 0,
+              end: 1,
+              count: 1,
+            },
+          },
+        },
+      })
+
+      // 4. Wait for workflow to complete
+      console.log(
+        `Submitted E2E Video Screenshot Workflow Task. ID: ${task.id}. Awaiting completion...`,
+      )
+      const completedTask = await workflowService.executeWait(task, 45000)
+
+      // 5. Verification
+      expect(completedTask.status).toBe('completed')
+
+      // Verify task output contains list of screenshots
+      const output = completedTask.output as unknown as { screenshots: { key: string }[] }
+      expect(output).toBeDefined()
+      expect(output.screenshots).toBeDefined()
+      expect(output.screenshots.length).toBe(1)
+      expect(output.screenshots[0].key).toContain('screenshots/')
+    }, 50000)
+  },
+)
